+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added rendering support for selecting texture coordinate attributes to xc3_wgpu.
 * Added support for morph target names to glTF export.
 * Added support for isolating color channels in xc3_viewer with the RGBA or XYZW keys.
+* Added support to xc3_tex for exporting cube maps with the `--cube` flag and 3D textures with the `--depth` flags.
 
 ### Changed
 * Changed shader database to store texture layering information for each output and channel.